@@ -4,9 +4,7 @@ from conans import ConanFile, CMake, tools
 
 class HomestoreConan(ConanFile):
     name = "homestore"
-
-    version = "2.1.11"
-
+    version = "2.1.12"
     revision_mode = "scm"
 
     license = "Proprietary"
