@@ -63,8 +63,8 @@ struct dev_info {
 #endif
 
 #define HOMESTORE_LOG_MODS                                                                                             \
-    btree_structures, btree_nodes, btree_generics, cache, device, httpserver_lmod, varsize_blk_alloc, volume, flip,    \
-        cp, metablk, indx_mgr, logstore, IOMGR_LOG_MODS
+    btree_structures, btree_nodes, btree_generics, cache, device, httpserver_lmod, blkalloc, volume, flip, cp,         \
+        metablk, indx_mgr, logstore, IOMGR_LOG_MODS
 
 template < typename T >
 std::string to_hex(T i) {
