@@ -47,7 +47,7 @@ using log_level = spdlog::level::level_enum;
 SDS_LOGGING_INIT(HOMESTORE_LOG_MODS)
 
 // SDS_LOGGING_INIT(cache_vmod_evict, cache_vmod_write, iomgr, btree_structures, btree_nodes, btree_generics,
-//                 varsize_blk_alloc, VMOD_VOL_MAPPING, httpserver_lmod, volume)
+//                 blkalloc, VMOD_VOL_MAPPING, httpserver_lmod, volume)
 
 /*** CLI options ***/
 SDS_OPTION_GROUP(
